rozpoczęcie pracy nad nowym presetem
